requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jira==3.5.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.4
plotly==5.17.0
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Frontend dependencies
streamlit==1.28.1
requests==2.31.0
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4
streamlit-authenticator==0.2.3
openai==1.3.7
O365==2.0.25 